pyproject.toml
[tool.poetry]
name = "rotab"
version = "0.1.0"
description = "ROTAB: a template that moves with your thinking."
packages = [{ include = "rotab", from = "src" }]
authors = ["yuta.itatsu <yuta.itatsu@gmail.com>"]
readme = "README.md"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
license = "MIT"
homepage = "https://github.com/project-ro/rotab"
repository = "https://github.com/project-ro/rotab"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pandas = ">=1.5.0,<3.0.0"
pyyaml = ">=6.0.0,<7.0.0"
pydantic = ">=2.10.0,<3.0.0"
typing_extensions = "^4.0"
typer = ">=0.4.0,<1.0.0"
polars = "^1.31.0"
fsspec = "^2025.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
tox = "^3.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rotab = "rotab.cli.cli:main"
